FourthHomework/src/main/java/task7/dao/HumanDaoImpl.java
package task7.dao;

import org.hibernate.Session;
import org.hibernate.SessionFactory;
import org.hibernate.Transaction;
import task7.model.Human;
import util.MysqlSessionFactory;

public class HumanDaoImpl implements HumanDao {
    private final SessionFactory sessionFactory;

    public HumanDaoImpl() {
        this(MysqlSessionFactory.getInstance());
    }

    public HumanDaoImpl(SessionFactory sessionFactory) {
        this.sessionFactory = sessionFactory;
    }

    @Override
    public void create(Human human) {
        Transaction tx = null;
        try (Session session = sessionFactory.openSession()) {
            tx = session.beginTransaction();
            session.saveOrUpdate(human);
            tx.commit();
        } catch (Exception e) {
            if (tx != null) tx.rollback();
        }
    }

    @Override
    public Human findById(Integer id) {
        Human human = null;
        Transaction tx = null;
        try (Session session = sessionFactory.openSession()) {
            tx = session.beginTransaction();
            human = session.get(Human.class, id);
            session.getTransaction();
            tx.commit();
        } catch (Exception e) {
            if (tx != null) tx.rollback();
        }
        return human;
    }

    @Override
    public void update(Human human) {
        create(human);
    }

    @Override
    public void delete(Human human) {
        Transaction tx = null;
        try (Session session = sessionFactory.openSession()) {
            tx = session.beginTransaction();
            session.delete(human);
            session.getTransaction();
            tx.commit();
        } catch (Exception e) {
            if (tx != null) tx.rollback();
        }
    }
}
